requirements.txt
python-dotenv
pandas
numpy
pymongo>=4.0.0
certifi
scikit-learn
dill
pyaml
mlflow

##-e .